Relocated LocalDataBase code
