+++ lib/models/dateAndCalory.dart
@@ -1,0 +1,12 @@
+class DateAndCalory {
+  int id;
+  DateTime date;
+  double calory;
+
+  DateAndCalory({
+    this.id,
+    this.date,
+    this.calory,
+  });
+  
+}
