+++ lib/pages/auth.dart
@@ -1,9 +1,9 @@
+import 'package:calory_calc/providers/local_providers/userProvider.dart';
 import 'package:calory_calc/utils/adClickHelper.dart';
 import 'package:flutter/material.dart';
 import 'package:calory_calc/design/theme.dart';
 
 import 'package:calory_calc/models/dbModels.dart';
-import 'package:calory_calc/utils/databaseHelper.dart';
 import 'package:calory_calc/utils/dataLoader.dart';
 
 import 'package:gradient_widgets/gradient_widgets.dart';
