+++ lib/pages/authLastPage.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:calory_calc/models/dbModels.dart';
 import 'package:calory_calc/utils/adClickHelper.dart';
-import 'package:calory_calc/utils/databaseHelper.dart';
+import 'package:calory_calc/providers/local_providers/userProvider.dart';
 import 'package:calory_calc/widgets/customRadioSelector.dart';
 
 class ActivitiSelectPage extends StatefulWidget {
